+++ mobile/lib/pages/home.dart
@@ -1,0 +1,150 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import '../widgets/bottomnav.dart';
+
+class HomeScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Trill'),
+        leading: IconButton(
+          icon: Icon(Icons.menu),
+          onPressed: () {
+            // Open sidebar
+          },
+        ),
+        actions: [
+          IconButton(
+            icon: Icon(Icons.search),
+            onPressed: () {
+              final model = Provider.of<TrillBottomNavigatorModel>(context,
+                  listen: false);
+              model.setSelectedIndex(1);
+              // Open Search Menu
+              Navigator.pushNamed(context, '/search');
+            },
+          )
+        ],
+      ),
+      body: Container(
+        padding: const EdgeInsets.all(16.0),
+        width: double.infinity,
+        child: Expanded(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                'Hello, username!',
+                style: TextStyle(
+                  color: Colors.blue,
+                  fontSize: 20,
+                  fontWeight: FontWeight.bold,
+                  fontFamily: 'Roboto',
+                ),
+              ),
+              SizedBox(height: 20),
+              Text(
+                'Popular Albums This Month',
+                textDirection: TextDirection.ltr,
+                style: TextStyle(
+                  color: Colors.white,
+                  fontSize: 12,
+                  fontWeight: FontWeight.bold,
+                  fontFamily: 'Roboto',
+                ),
+              ),
+              SizedBox(height: 10),
+              // Image links will go here
+              SizedBox(height: 20),
+              Text(
+                'Popular Albums Among Friends',
+                textDirection: TextDirection.ltr,
+                style: TextStyle(
+                  color: Colors.white,
+                  fontSize: 12,
+                  fontWeight: FontWeight.bold,
+                  fontFamily: 'Roboto',
+                ),
+              ),
+              SizedBox(height: 10),
+              // Image links will go here
+              SizedBox(height: 20),
+              Text(
+                'Recent Friends Reviews',
+                textDirection: TextDirection.ltr,
+                style: TextStyle(
+                  color: Colors.white,
+                  fontSize: 12,
+                  fontWeight: FontWeight.bold,
+                  fontFamily: 'Roboto',
+                ),
+              ),
+              SizedBox(height: 10),
+              Container(
+                decoration: BoxDecoration(
+                  border: Border.all(
+                    color: Colors.black38,
+                  ),
+                  color: Color(0x1f989696),
+                ),
+                padding: const EdgeInsets.all(16),
+                child: Column(
+                  children: [
+                    Text(
+                      '"I wasn\'t a fan at first, but now I can\'t stop listening!"',
+                      style: TextStyle(
+                        color: Colors.white,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              SizedBox(height: 10),
+              Container(
+                decoration: BoxDecoration(
+                  border: Border.all(
+                    color: Colors.black38,
+                  ),
+                  color: Color(0x1f989696),
+                ),
+                padding: const EdgeInsets.all(16),
+                child: Column(
+                  children: [
+                    Text(
+                      '"This album is definitely worth checking out!"',
+                      style: TextStyle(
+                        color: Colors.white,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              SizedBox(height: 10),
+              Container(
+                decoration: BoxDecoration(
+                  border: Border.all(
+                    color: Colors.black38,
+                  ),
+                  color: Color(0x1f989696),
+                ),
+                padding: const EdgeInsets.all(16),
+                child: Column(
+                  children: [
+                    Text(
+                      '"My name is Wesley Wales Anderson and if I were a music album, this would be me!"',
+                      style: TextStyle(
+                        color: Colors.white,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+      bottomNavigationBar: TrillBottomNavigatorState(),
+    );
+  }
+}
