+++ mobile/lib/pages/profile.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import '../widgets/bottomnav.dart';
+
+class ProfileScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Username\'s Profile'),
+      ),
+      bottomNavigationBar: TrillBottomNavigatorState(),
+    );
+  }
+}
