+++ mobile/lib/pages/signup.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+
+class SignUpScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Sign Up'),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Image.asset('images/logo.png'),
+            SizedBox(height: 30),
+            Text(
+              'Create an account to continue.',
+              style: TextStyle(
+                fontSize: 10,
+              ),
+            ),
+            SizedBox(height: 10),
+            TextField(
+              decoration: InputDecoration(
+                hintText: 'Username',
+              ),
+            ),
+            SizedBox(height: 10),
+            TextField(
+              decoration: InputDecoration(
+                hintText: 'Email',
+              ),
+            ),
+            SizedBox(height: 10),
+            TextField(
+              decoration: InputDecoration(
+                hintText: 'Password',
+              ),
+            ),
+            SizedBox(height: 10),
+            ElevatedButton(
+              child: Text('Sign Up'),
+              onPressed: () {
+                // Sign the user up
+                Navigator.pushNamed(context, '/');
+              },
+            ),
+            SizedBox(height: 10),
+            Text(
+              'Already have an account? Head to the login page.',
+              style: TextStyle(
+                fontSize: 10,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
